requirements.txt
fastapi
uvicorn
python-dotenv
requests
google-auth
google-auth-oauthlib
google-auth-httplib2
google-cloud-aiplatform
vertexai
PyJWT
beanie
motor
pydantic
email-validator
aiohttp